# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Config Manager API
  version: 0.1.0
  description: |
    High-assurance Configuration Management System API

    Built with Effect-TS and strict functional programming principles.
  contact:
    name: API Support
    url: https://github.com/your-org/config-manager
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:3000/api
    description: Nuxt server API routes

tags:
  - name: feature-flags
    description: Feature flag management operations

paths:
  /api/flags:
    post:
      operationId: createFeatureFlag
      summary: Create a new feature flag
      description: |
        Creates a new feature flag with validation:
        - Key must be kebab-case (lowercase alphanumeric with hyphens)
        - Description must be non-empty and max 500 characters
        - Key must be unique
      tags:
        - feature-flags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFlagInput"
            examples:
              darkMode:
                summary: Dark mode feature flag
                value:
                  key: dark-mode
                  defaultValue: false
                  description: Enable dark mode theme for users
              premiumFeature:
                summary: Premium feature
                value:
                  key: premium-analytics
                  defaultValue: true
                  description: Advanced analytics for premium users
      responses:
        "201":
          description: Feature flag created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlag"
              examples:
                success:
                  value:
                    key: dark-mode
                    defaultValue: false
                    description: Enable dark mode theme for users
                    createdAt: "2025-01-31T12:00:00.000Z"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              examples:
                invalidKey:
                  summary: Invalid key format
                  value:
                    error: InvalidFlagKeyError
                    message: Key must be kebab-case lowercase alphanumeric characters separated by hyphens
                emptyDescription:
                  summary: Empty description
                  value:
                    error: InvalidFlagDescriptionError
                    message: Description cannot be empty
                descriptionTooLong:
                  summary: Description too long
                  value:
                    error: InvalidFlagDescriptionError
                    message: Description cannot exceed 500 characters
        "409":
          description: Flag already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
              examples:
                duplicate:
                  value:
                    error: FlagAlreadyExistsError
                    message: "Flag with key 'dark-mode' already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerError"

components:
  schemas:
    CreateFlagInput:
      type: object
      required:
        - key
        - defaultValue
        - description
      properties:
        key:
          type: string
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          minLength: 1
          description: Kebab-case identifier (lowercase alphanumeric with hyphens)
          example: dark-mode
        defaultValue:
          type: boolean
          description: Default value for the feature flag
          example: false
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: Human-readable description of the feature flag
          example: Enable dark mode theme for users
      additionalProperties: false

    FeatureFlag:
      type: object
      required:
        - key
        - defaultValue
        - description
        - createdAt
      properties:
        key:
          type: string
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          description: Kebab-case identifier
          example: dark-mode
        defaultValue:
          type: boolean
          description: Default value for the feature flag
          example: false
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: Human-readable description
          example: Enable dark mode theme for users
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp of creation
          example: "2025-01-31T12:00:00.000Z"
      additionalProperties: false

    ValidationError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - ValidationError
            - InvalidFlagKeyError
            - InvalidFlagDescriptionError
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional error details (for ValidationError only)
      additionalProperties: false

    ConflictError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - FlagAlreadyExistsError
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
      additionalProperties: false

    ServerError:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - FlagRepositoryError
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
      additionalProperties: false